AMC_BLDC: Alignment with the latest changes on MBD (#250)
* AMC_BLDC: Update generated code with features to prepare for multi CAN packet + transmit STATUS message

* AMC_BLDC: Update MBDAgent glue code to work with new generated code

* AMC_BLDC: Cleanup MBDAgent + Remove can-raw2struct folder (no more used)

* AMC_BLDC: MDBAgent update pole_pairs during initialization and runtime

* AMC_BLDC: During the periodic tick function and FOC_inner_loop use the ConfigurationParameters model's output instead of the initialization paraemtere. Update the type of poles parameter in pwm.h optimizing memory usage.

+++ emBODY/eBcode/arch-arm/board/amcbldc/application06/src/motorhal/pwm.h
@@ -49,7 +49,7 @@
 typedef struct
 {
     uint32_t    mode;
-    uint32_t    poles;
+    uint8_t    poles;
 } pwmConfTypeDef;
 
 // we keep int32_t even if the adc gets only int16_t values
